Address review feedback on device availability state PR
- Move sample_http_data, sample_mqtt_data, sample_offline_status_payload,
  and sample_online_status_payload fixtures to tests/conftest.py so they
  are shared across test_device.py and test_client.py
- Remove MQTT_AVAILABILITY_FRESHNESS_WINDOW from the public __all__ export
  as it is an internal implementation detail
- Add comment to const.py explaining the 6-minute freshness window value
  (5-min MQTT heartbeat + 1-min jitter buffer)
